+++ index.js
@@ -10,7 +10,6 @@ const plugin = {
         socketUrl={args.socketUrl}
         socketPath={args.socketPath}
         protocolOptions={args.protocolOptions}
-        interval={args.interval}
         initPayload={args.initPayload}
         title={args.title}
         subtitle={args.subtitle}
@@ -35,6 +34,10 @@ const plugin = {
         docViewer={args.docViewer}
         persistentMenu={args.persistentMenu}
         displayUnreadCount={args.displayUnreadCount}
+        showMessageDate={args.showMessageDate}
+        customMessageDelay={args.customMessageDelay}
+        tooltipPayload={args.tooltipPayload}
+        tooltipDelay={args.tooltipDelay}
       />, document.querySelector(args.selector)
     );
   }