+++ src/index.js
@@ -26,6 +26,10 @@ const ConnectedWidget = (props) => {
       this.onEvents = [];
     }
 
+    isInitialized() {
+      return this.socket !== null && this.socket.connected;
+    }
+
     on(event, callback) {
       if (!this.socket) {
         this.onEvents.push({ event, callback });
@@ -57,12 +61,7 @@ const ConnectedWidget = (props) => {
       this.onEvents.forEach((event) => {
         this.socket.on(event.event, event.callback);
       });
-      if (this.onSocketEvent.disconnect) {
-        this.socket.on('disconnect', this.onSocketEvent.disconnect);
-      }
-      if (this.onSocketEvent.connect) {
-        this.socket.on('connect', this.onSocketEvent.connect);
-      }
+
       this.onEvents = [];
       Object.keys(this.onSocketEvent).forEach((event) => {
         this.socket.on(event, this.onSocketEvent[event]);
@@ -93,7 +92,6 @@ const ConnectedWidget = (props) => {
   return (
     <Provider store={store}>
       <Widget
-        interval={props.interval}
         initPayload={props.initPayload}
         title={props.title}
         subtitle={props.subtitle}
@@ -117,6 +115,10 @@ const ConnectedWidget = (props) => {
         displayUnreadCount={props.displayUnreadCount}
         persistentMenu={props.persistentMenu}
         socket={sock}
+        showMessageDate={props.showMessageDate}
+        customMessageDelay={props.customMessageDelay}
+        tooltipPayload={props.tooltipPayload}
+        tooltipDelay={props.tooltipDelay}
       />
     </Provider>
   );
@@ -124,7 +126,6 @@ const ConnectedWidget = (props) => {
 
 ConnectedWidget.propTypes = {
   initPayload: PropTypes.string,
-  interval: PropTypes.number,
   title: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
   subtitle: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
   protocol: PropTypes.string,
@@ -153,13 +154,16 @@ ConnectedWidget.propTypes = {
   docViewer: PropTypes.bool,
   persistentMenu: PropTypes.objectOf(PropTypes.arrayOf(PropTypes.string)),
   customComponent: PropTypes.func,
-  displayUnreadCount: PropTypes.bool
+  displayUnreadCount: PropTypes.bool,
+  showMessageDate: PropTypes.oneOfType([PropTypes.bool, PropTypes.func]),
+  customMessageDelay: PropTypes.func,
+  tooltipPayload: PropTypes.string,
+  tooltipDelay: PropTypes.number
 };
 
 ConnectedWidget.defaultProps = {
   title: 'Welcome',
   customData: {},
-  interval: 2000,
   inputTextFieldHint: 'Type a message...',
   connectingText: 'Waiting for server...',
   fullScreenMode: false,
@@ -179,7 +183,16 @@ ConnectedWidget.defaultProps = {
   persistentMenu: [],
   showCloseButton: true,
   showFullScreenButton: false,
-  displayUnreadCount: false
+  displayUnreadCount: false,
+  showMessageDate: false,
+  customMessageDelay: (message) => {
+    let delay = message.length * 30;
+    if (delay > 2 * 1000) delay = 3 * 1000;
+    if (delay < 400) delay = 1000;
+    return delay;
+  },
+  tooltipPayload: null,
+  tooltipDelay: 500
 };
 
 export default ConnectedWidget;