+++ src/store/reducers/helper.js
@@ -3,14 +3,14 @@ import { MESSAGES_TYPES, MESSAGE_SENDER, SESSION_NAME } from 'constants';
 
 import { Video, Image, Message, Snippet, QuickReply } from 'messagesComponents';
 
-
 export function createNewMessage(text, sender) {
   return Map({
     type: MESSAGES_TYPES.TEXT,
     component: Message,
     text,
     sender,
-    showAvatar: sender === MESSAGE_SENDER.RESPONSE
+    showAvatar: sender === MESSAGE_SENDER.RESPONSE,
+    timestamp: new Date().getTime()
   });
 }
 
@@ -20,7 +20,8 @@ export function createLinkSnippet(elements, sender) {
     component: Snippet,
     elements: elements,
     sender,
-    showAvatar: true
+    showAvatar: true,
+    timestamp: new Date().getTime()
   });
 }
 
@@ -31,7 +32,8 @@ export function createVideoSnippet(video, sender) {
     title: video.title,
     video: video.video,
     sender,
-    showAvatar: true
+    showAvatar: true,
+    timestamp: new Date().getTime()
   });
 }
 
@@ -42,7 +44,8 @@ export function createImageSnippet(image, sender) {
     title: image.title,
     image: image.image,
     sender,
-    showAvatar: true
+    showAvatar: true,
+    timestamp: new Date().getTime()
   });
 }
 
@@ -55,7 +58,8 @@ export function createQuickReply(quickReply, sender) {
     quick_replies: List(quickReply.quick_replies),
     sender,
     showAvatar: true,
-    chosenReply: null
+    chosenReply: null,
+    timestamp: new Date().getTime()
   });
 }
 
@@ -65,7 +69,8 @@ export function createComponentMessage(component, props, showAvatar) {
     component,
     props,
     sender: MESSAGE_SENDER.RESPONSE,
-    showAvatar
+    showAvatar,
+    timestamp: new Date().getTime()
   });
 }