+++ CHANGELOG.md
@@ -2,6 +2,61 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.7.2](https://https///compare/v0.7.1...v0.7.2) (2019-12-04)
+
+
+### Features
+
+* add disconnect reason in socketjs LEAVE ([290e356](https://https///commit/290e356))
+
+### [0.7.1](https://https///compare/v0.7.0...v0.7.1) (2019-11-27)
+
+## [0.7.0](https://https///compare/v0.6.6...v0.7.0) (2019-11-21)
+
+
+### Bug Fixes
+
+* init payload is now sent only when opened ([efebacb](https://https///commit/efebacb))
+
+
+### Features
+
+* added a prop tooltipDelay ([8c8cd95](https://https///commit/8c8cd95))
+* smart typing delay and typing indication ([8738c63](https://https///commit/8738c63))
+* tooltip popup ([04e1279](https://https///commit/04e1279))
+
+### [0.6.6](https://https///compare/v0.6.5...v0.6.6) (2019-10-30)
+
+
+### Bug Fixes
+
+* margin 0 is prefered to margin 0px ([2c0ca86](https://https///commit/2c0ca86))
+* no margin on last child only applies to <p> ([4fe3e9b](https://https///commit/4fe3e9b))
+
+
+### Features
+
+* add margin to <p> to display newline ([c7df04b](https://https///commit/c7df04b))
+
+### [0.6.5](https://https///compare/v0.6.4...v0.6.5) (2019-10-25)
+
+
+### Bug Fixes
+
+* duplicate registration of connect and disconnect event ([263b847](https://https///commit/263b847))
+* duplicate user connection and small refactoring ([58e8925](https://https///commit/58e8925))
+* function binding ([0a035c0](https://https///commit/0a035c0))
+* invalid date if timestamp not present ([4c7be48](https://https///commit/4c7be48))
+* link and add container-class to quickreplies ([67f7fcf](https://https///commit/67f7fcf))
+* lost of path in sockjs implementation ([46adcab](https://https///commit/46adcab))
+* message's date CSS and removed seconds from date ([e0fe602](https://https///commit/e0fe602))
+* sockjs duplicate connection ([991cc6a](https://https///commit/991cc6a))
+
+
+### Features
+
+* display message's date on chat ([3fef7c8](https://https///commit/3fef7c8))
+
 ### [0.6.4](https://https///compare/v0.6.3...v0.6.4) (2019-10-18)